+++ app/src/main/java/dk/scuffed/whiteboardapp/pipeline/stages/GaussianBlurStage.kt
@@ -3,7 +3,7 @@ package dk.scuffed.whiteboardapp.pipeline.stages
 import android.content.Context
 import android.opengl.GLES20
 import dk.scuffed.whiteboardapp.R
-import dk.scuffed.whiteboardapp.openGL.*
+import dk.scuffed.whiteboardapp.opengl.*
 import dk.scuffed.whiteboardapp.pipeline.FramebufferInfo
 import dk.scuffed.whiteboardapp.pipeline.GLOutputStage
 import dk.scuffed.whiteboardapp.pipeline.Pipeline
