+++ app/src/main/java/dk/scuffed/whiteboardapp/pipeline/stages/SegmentationStage.kt
@@ -2,12 +2,11 @@ package dk.scuffed.whiteboardapp.pipeline.stages
 
 import android.content.Context
 import android.graphics.Bitmap
-import android.util.Log
 import android.util.Size
 import dk.scuffed.whiteboardapp.pipeline.*
 import dk.scuffed.whiteboardapp.segmentation.PPSegmentation
 
-internal class SegmentationStage(context: Context, segmentationModel: PPSegmentation.Model,  private val inputBitmap: Bitmap, private val pipeline: Pipeline)
+internal class SegmentationStage(context: Context, segmentationModel: PPSegmentation.Model,  private val inputBitmap: Bitmap, pipeline: Pipeline)
     : BitmapOutputStage(pipeline, Size(inputBitmap.width, inputBitmap.height), inputBitmap.config) {
 
     private var segmentor: PPSegmentation
