+++ app/src/main/java/dk/scuffed/whiteboardapp/pipeline/stages/SobelStage.kt
@@ -3,10 +3,7 @@ package dk.scuffed.whiteboardapp.pipeline.stages
 import android.content.Context
 import android.opengl.GLES20
 import dk.scuffed.whiteboardapp.R
-import dk.scuffed.whiteboardapp.openGL.glActiveTexture
-import dk.scuffed.whiteboardapp.openGL.glBindTexture
-import dk.scuffed.whiteboardapp.openGL.glGetUniformLocation
-import dk.scuffed.whiteboardapp.openGL.glUniform1i
+import dk.scuffed.whiteboardapp.opengl.*
 import dk.scuffed.whiteboardapp.pipeline.FramebufferInfo
 import dk.scuffed.whiteboardapp.pipeline.GLOutputStage
 import dk.scuffed.whiteboardapp.pipeline.Pipeline
