+++ app/src/main/java/dk/scuffed/whiteboardapp/pipeline/stages/TextureStage.kt
@@ -2,11 +2,10 @@ package dk.scuffed.whiteboardapp.pipeline.stages
 
 import android.content.Context
 import android.graphics.Bitmap
-import android.graphics.BitmapFactory
 import android.opengl.GLES20
 import android.util.Size
 import dk.scuffed.whiteboardapp.R
-import dk.scuffed.whiteboardapp.openGL.*
+import dk.scuffed.whiteboardapp.opengl.*
 import dk.scuffed.whiteboardapp.pipeline.*
 import java.nio.ByteBuffer
